requirements.txt
numpy
pandas
matplotlib
opencv-python-headless
torch
torchvision
scikit-learn
streamlit
Pillow